Fix grade loop var, split verdict schema, link Azure docs
